feat: Add `columns` API query parameter to filter table columns (#1829)
Implements https://github.com/electric-sql/electric/issues/1804


### Overview
- I've picked `columns` as the query parameter name, as `select` might
confuse people and get compared to SQL `SELECT`s, and `filter` is too
close to what the `where` parameter does.
- The columns are comma separated column names, parsed at the validation
stage
- If the parameter is specified, it *has to include* the primary key
columns
- We could make it such that the PK is always included, but it feels
like it might be confusing - explicit is better?
- For tables without a primary key, we treat _all_ columns as primary
keys - we might want to revisit that w.r.t. to this feature
- Columns are validated also by ensuring that all specified column names
match the ones in the schema
- The selected columns are stored in the `Shape` definition as an array
of column names
- We had the choice of potentially filtering the `table_info` data based
on the selection of columns instead of storing them separately, but this
might cause issues with relation change handling and generally I prefer
the definition of the shape to contain all information related to it
(i.e. the schema at the time of creation + what column filters where
applied, rather than compacting the two)
- I modified `Shape` related APIs for converting changes etc to also
apply the column filtering - this is all we need to ensure it works
correctly for replication stream log entries
- In the `Snapshotter`, I modified a `get_column_names` method to also
apply the filtering if present, which takes care of the snapshot log
entries.

### Other changes
- I've changed `Shapes.new` to return the errors along with the field
they are associated with (`:root_table` or `:where` or `:columns`) in
order to return correct validation errors, since all of the stronger
validation occrus at `cast_root_table` when the shape is created and PG
is used but it really validates more than just `root_table`.
- I've updated the client to accept a `columns` argument which is a list
of strings
- I've updated the OpenAPI spec to include the `columns` parameter

### Things to consider
- How do we want to handle column names with special names (quoted)? In
my opinion the client needs to type them exactly as they are on
postgres, otherwise they get a 400 validation error back telling them
which columns are invalid, so it should be fairly easy to fix.
- Replication publication filtering of columns is available, but
[updating it might cause
errors](https://www.postgresql.org/docs/15/logical-replication-col-lists.html#LOGICAL-REPLICATION-COL-LIST-COMBINING)
according to the PG docs, so I'm not sure if that's something we want to
consider