chore(sync-service): Set default log level to 'info' in production (#1855)
We have detailed debug logs that log every transaction incoming from
Postgres, including its every change. This is not a good default for a
production environment where the rate of writes may easily spam the
logs, drawing them almost useless.

We leave debug as the default for the dev env.