chore(sync-service): Decouple Timeline from ShapeCache
Move shape purging from Timeline to ShapeCache where it belongs

This change reduces coupling between Timeline and ShapeCache modules,
resulting in easier comprehension and maintenance of this small part of
the code.

ConnectionManager serves as the supervisor of the connection startup
procedure and that's where we keep the knowledge about how a change in
PG timeline affects the shape cache.

As a bonus, we don't have to first start all shape consumers just to
bring them all down afterwards, as we've been doing up until this
change.