Simplify configuration of the electric app
When electric is included as a dependency then the configuration done in
our `config/runtime.exs` is not used, and the application needs to set
electric up itself.

Because we've made every configuration option required in the code
(relying on runtime.exs) this is onerous.

But, because every option except the db connection actually has sensible
defaults, we can simplify this and reduce the required configuration
settings down to the db connection stuff.

The only downside to this is that now we have default values for the
parameters in two locations, runtime.exs and with every
`Application.get_env/3` call.

I took the opportunity to re-factor the configuration stuff and move it
out of the application.ex module.