Add parameter to API to send entire row on update (#1807)
With `replica=full` update messages include the full row, not just
the changed columns.

Phoenix.LiveView Stream integration requires the full row to perform an
update, not just the changed values. This adds a parameter to the shape
definition that turns off the column diffing for updates.

Since we were just dropping unchanged values this is a minor change to
the wiring. Most of the work here is adding the configuration.