docs: deployment guide, with supporting config and integrations. (#1836)
Fleshing out deployment guide. To support it, added a sync service
config page in the api docs section and specific integrations pages for
various deployment platforms.

Quite a lot to review here. Check out the:

- [Sync service config page at
/docs/api/config](https://deploy-preview-1836--electric-next.netlify.app/docs/api/config)
- [Updated Typescript client page at
/docs/api/clients/typescript](https://deploy-preview-1836--electric-next.netlify.app/docs/api/clients/typescript)
- [Deployment page at
/docs/guides/deployment](https://deploy-preview-1836--electric-next.netlify.app/docs/guides/deployment)
- Individual integration pages in the "Integrations" section of the docs
nav, e.g.:
[AWS](https://deploy-preview-1836--electric-next.netlify.app/docs/integrations/aws),
[React](https://deploy-preview-1836--electric-next.netlify.app/docs/integrations/react)
and
[Supabase](https://deploy-preview-1836--electric-next.netlify.app/docs/integrations/supabase)

Note that:

- Netlify has a one-click-deploy form that uses
https://github.com/electric-sql/standalone-basic-example
-
[Render](https://deploy-preview-1836--electric-next.netlify.app/docs/integrations/supabase)
has a [Blueprint repo](https://github.com/electric-sql/render-blueprint)
that also uses the same standalone-basic-example

As @KyleAMathews commented on Discord, these might be better to include
in the monorepo and update using a GitHub action. Perhaps that can be a
subsequent Issues / PR.

The integrations also include a [Phoenix
page](https://deploy-preview-1836--electric-next.netlify.app/docs/integrations/phoenix)
which is based on merging #1833 and publishing an updated release for
https://hexdocs.pm/electric_phoenix/0.1.0-dev-2/Electric.Phoenix.html

Note that the deployment docs all assume a single `DATABASE_URL`, i.e.:
this is pre multi-tenant.

---------

Co-authored-by: Kyle Mathews <mathews.kyle@gmail.com>