feat (sync-service): Prevent shape consumer errors from affecting other shapes (#2009)
Fixes #1925 . Errors that occur while consuming the replication stream
that are to do with a specific shape, cause that shape's consumer to
remove the shape and shut down, leaving the other shapes unaffected.

Errors can occur:
1. In the selector which happens on a process common to all shapes. 
2. On the consumer process.

This PR addresses both types of error. Previous to this PR, these errors
would cause the sync service to get into an infinite crash loop and
would stop responding to HTTP requests.