feat: enable error handler to recover from failed HTTP requests (#2011)
This PR modifies the `onError` handler that is supported by the shape
stream such that it can recover from certain HTTP request errors. For
example, if the HTTP request fails with a 401 because of an expired
token, the handler could fetch a new token, and try again by returning
the modified HTTP headers. The shapestream will then retry.