examples: update gatekeeper auth client to use `onError`. (#2028)
This is what I requested in
https://github.com/electric-sql/electric/pull/2018#issuecomment-2491085763

The client now uses the new onError to handle 401/403 and reconnect with
a new auth token.

One observation: with the onError being defined before the shape is
instantiated, I'm not sure what the best control flow is to handle a
`405` must refetch. Certainly it's hard to handle in the onError
callback. I guess this is what's being discussed in
https://github.com/electric-sql/electric/issues/1997 etc.