fix: Clean up shape folders and remove unrecoverable shapes (#2051)
Followup to https://github.com/electric-sql/electric/pull/2049

We are currently leaving folders behind when deleting shapes, which do
take up a little bit of space and eventually could be a bit of an issue.
This PR takes care of removing everything.

Additionally, when a shape fails to recover by not being able to start
the consumer at all (see CubDB issue), the shape data is removed using a
new method called `unsafe_cleanup!` which simply `rm -rf`s the shape
directory. Not great to use while the shape is active as it can cause
various bugs, but great for cleaning up.