examples: always start clean and update running instructions. (#2045)
I've been working with the examples quite a lot and I've commonly
experienced stale database and disk state from running `docker compose
up` in the foreground followed by `ctl+c` to stop and also from
forgetting to run `pnpm backend:down`.

So ... as a systemic solution, I've updated the `example-backend:up`
script in the root package.json to always run `example-backend:down`
first. This means that examples always start with a clean slate and no
weird / stuck table and or disk state.

I also then did a pass through the example READMEs, explaining this and
re-formatting the instructions a touch.