chore (sync service): setup sentry (#2099)
This PR configures Electric to track errors with sentry.
To enable sentry for the single tenant Electric app you need to pass a
`SENTRY_DSN` env var.

I've tested this with the cloud deploy and errors are tracked
successfully in sentry.

TODO:
- [ ] add stack_id and request_id meta data to the recorded error events
(*)

(*) The easy way is to configure sentry to add the `stack_id` and
`request_id` meta data to the event context. That way it will show up in
the error's details but it won't be searchable filterable on those
fields! So a better way would be to add that meta data as tags. But not
sure if that is configurable out of the box, might need to write a
custom LoggerHandler.

Adding the meta data as tags is possible by extending all
`Logger.metadata` calls (from @icehaunter's recent PR) with a `sentry:
[tags: %{stack_id: stack_id, shape_handle: shape_handle}]` key. Cf.
https://github.com/getsentry/sentry-elixir/blob/master/lib/sentry/logger_handler.ex#L205
for more info.

---------

Co-authored-by: Ilia Borovitinov <ilia@electric-sql.com>