Add partitioned tables support (#2197)
Allows for subscribing to partitions of a partitioned table (which you
could already do) but also subscribing to the root, partitioned, table
and receiving all updates across all partitions.

Fixes #2118 

Major changes:

- Add support for partitioned table types in the various inspector
queries
- Include partition hierarchy information in relation information from
pg and in inspector relation info
- Force reload of table information when receiving a new relation that
is a partition
- Add partition information to `Filter` so that it can correctly pass on
a change on a partition table to a shape on the partition root.
- For a shape subscribed to the partition root, changes coming in on
partitions will be re-written so that their relation matches the
partition root. The change `key`s are not being re-written - I don't
think this is an issue as the keys are guaranteed to be unique and
that's enough.
- Relation messages notifying of a new partition table are forwarded
onto root-partition consumers so the shape can be updated with knowledge
of the new partition (this does not cause the shape to be deleted)
- I've added a full-stack test to verify this stuff so that we're not
making assumptions about pg's handling of partitions

Minor:

- support querying the inspector based on the {schema, name} relation
tuple, as well as the table name string because I needed to be able to
look up table info based
- Filter instances now have an inspector instance so they can do table
info lookups to get partition information
- The filter functions used by the dispatcher now return `{filter,
shape_ids}` so that the filter's state can be updated
- I've tweaked StubInspector so we can return richer info from the
load_relation call
- Added missing serialisation of a shape's replica setting