Use `unsafe_cleanup!/1` to delete data after consumer has stopped (#2271)
When cleaning up a shape we call `cleanup!` and (in the case of
FileStorage) attempt to remove the data directory while the CubDb
process is still running.

This fails (`File.rm_rf/1` returns `{:error, :eexist, _}`) because the
files are still in use.

Instead of removing the directory on `cleanup!` while the process(es)
are still running, instead teardown the consumer supervision tree and
use `unsafe_cleanup!/1` do remove the directories.

`unsafe_cleanup!` is already documented as something that should be
runnable without any active processes, so this is an ideal use.

Fixes problems with https://github.com/electric-sql/stratovolt/pull/243