feat(sync-service): Connect shape consumer spans to replication traces and add OTEL metrics (#2288)
This PR improves observability by:
- Connecting shape consumer spans to replication traces
- Adding `electric.shapes.total_shapes.count` metric to OTEL metrics
(Closes #2269 although I have not added "shapes that have an active
subscriber" as this information is already available in the traces)
- Adding `electric.storage.used` metric to OTEL metrics
- Adding CPU % utilisation per core telemetry and exposing them to the
OTEL metrics. CPU metrics are already available, but % utilisation per
core seems a simple one to comprehend and in a form our customers will
understand.