feat!: remove `offset` and add `lsn` header on the wire (#2294)
Closes #2214 

Removes `offset` property from all messages going forward. This
simplifies the wire protocol as currently we had discrepancies between
in-item offsets and header offsets when working with chunked initial
snapshot. Now headers are the only correct way to work with the offsets.
This is reflected in client libraries.

Adds two headers to *live* (i.e. non-snapshot, not-compacted) changes:
`lsn` and `op_position`. The `lsn` marks items in the same transaction
and provides a monotonic system-wide value for common prefix
calculations. It will be heavily used in further PRs. The `op_position`
lists an operation's order in the original transaction, allowing correct
operation ordering when merging and applying 2 shape streams together
without a need for a topological sort.