chore: Minor `Connection.Manager` and `Consumer` cleanup and flake reduction (#2305)
- Removes `ConnectionBackoff` struct, scope is small and it's verbose
(@robacourt nits are good we like them)
- Uses the `State` struct everywhere - since we have it it makes sense
to use it
- Remove `ShapeCache` dependency in `Consumer` as it already has
`ShapeStatus` available
- Uses `Consumer.whereis` in the `Consumer` tests for allowing Mox
assertions for consistency
- Fixes the `Consumer` test flakes by waiting for the
`set_snapshot_started` in the setup before running the tests, as the
tests "override" some Mox expectations and allowances set in the setup
if the snapshot is not ready before they get defined.

One thing that really worked for me that I hadn't tried before, is to
use a tool like `stress` (e.g. `stress -c 10`) to max out the resources
used by my machine while running tests to simulate a lower resource
system that the tests on CI run on. I managed to consistently reproduce
the consumer test flakes that way and debug them.


No changeset cause this is all refactoring and test fixing