fix: Fix total retry time from reconnnection attempts (#2306)
The sign of the total retry time was reversed. Lack of tests -> dumb
mistakes.

Is there any reason we don't have a test suite for the
`Connection.Manager`? It's a bit of an overhead to set it up, I've
separated the `ConnectionBackoff` logic to basically just slightly
enhance `:backoff` with total retry time and reverted the rest of the
logic in the connection manager to be how it was (probably was the best
option to start with).