Fix file corruption in Electric.Utils.external_merge_sort()
When the input file is small enough to fit into a single chunk. The
problem is that File.stream!, by default, looks for line break
characters and normalizes any \r\n into \n when reading from the file.
So the data it writes into the target file is different from the one it
read from chunk_0 if it so happens that the latter has the sequence of
bytes <<13, 10>> in it.

The issue was reliably reproduced by running

    mix test test/electric/utils_test.exs:30 --repeat-until-failure 1000

Co-authored-by: Garry Hill <garry@electric-sql.com>