Introduce Utils.write_stream_to_file!() for convenience
By default, File.stream!() scans the input file for line break
characters and normalizes \r\n to just \n. This makes it unsuitable for
copying or concatenating file contents with default options.

This change introduces a wrapper for the use cases of File.stream!()
that write to a file, making it easier to search the codebase for read
use cases and replace/fix them.