Fix file corruption in Electric.Utils.external_merge_sort() (#2350)
When the input file is small enough to fit into a single chunk. The
problem is that `File.stream!`, by default, looks for line break
characters and normalizes any `\r\n` into `\n` when reading from the
file. So the data it writes into the target file is different from the
one it read from chunk_0 if it so happens that the latter has the
sequence of bytes `<<13, 10>>` in it.

---------

Co-authored-by: Garry Hill <garry@electric-sql.com>
Co-authored-by: Ilia Borovitinov <ilia@electric-sql.com>