fix: Avoid long-lived listeners for changes due to connection process reuse (#2412)
Closes https://github.com/electric-sql/electric/issues/2289

It seems that connections are being handled by a process pool of sorts,
which means that they are never removed from the registry.

>Within a Bandit server, an HTTP/1 connection is modeled as a single
process. This process is tied to the lifecycle of the underlying TCP
connection; **in the case of an HTTP client which makes use of HTTP's
keep-alive feature** to make multiple requests on the same connection,
**all of these requests will be serviced by this same process**.

This is from [Bandit's
docs](https://github.com/mtrudel/bandit/blob/main/lib/bandit/http1/README.md)
and I've verified it by reproducing @magnetised 's situation described
in the issue.

I initially used an approach where an ephemeral process/`Task` is the
one registering for notifications and forwards them, so that it gets
automatically cleaned up by the registry, but I ended up using
`Registry.unregister` - it just gets a bit more complicated to ensure
that we unregister in all code paths for the request, since it might
exit early, and the registration occurs on the `validate` step of the
API, and the cleanup on the separate `serve_shape_log` step.

So either we spawn an ephemeral process and let it get cleaned up
automatically, or we have to ensure the cleanup is called despite it
being split across API calls. Perhaps we would need a `cleanup` API
method?