fix: Return chunked responses to live requesters (#2419)
Found out that live requesters would get the _entire_ transaction
regardless of its size while I was working on making the TS tests less
flaky.

I've modified the TS chunk test to test chunking both of the initial
snapshot and subsequent transactions, without arbitrary `sleep` calls,
so it was ending up doing a live request and noticed I was getting a
single >40kb response unless I "slept" a bit before requesting such that
it was not a live request.


The required change is really small, we just look for the chunk log
offset once we receive new changes.