feat: allow parameters in shape definitions (#2421)
Adds support for parameters in shape definitions on the server. API now
allows where clauses to have `$1` placeholders, which are replaced by
corresponding `params[1]=value` query parameters.

Query parameters are not used when either executing the initial query or
adding the clause to the publication filter, instead the value is
replaced in the original query after type-checking the usage. For
example, in `$1 > 2` clause the `$1` is treated as unknown, so the
overload resolver picks the `int4 > int4` operator, and checks if the
value passed as param 1 can be cast to desired type. If it can, then we
don't do de-serialization, and instead just do input escaping with an
explicit typecast, as Postgres is efficient enough to only cast const
values once. So the original query will be replaced with (for
`params[1]=3` query string) `'3'::int4 > 2`, where interpolation is
correctly escaped to avoid SQL injection.

One side-effect of this is that interpolation is done via rebuilding the
query from protobuf via PostgreSQL C Code, which normalizes the casing
and whitespace.

Closes #2295