feat: Implement separate, potentially pooled `ELECTRIC_QUERY_DATABASE_URL` (#2438)
Closes https://github.com/electric-sql/electric/issues/1885

- Introduce separate env var `ELECTRIC_QUERY_DATABASE_URL`
- Add another `connection_opts` field to `replication_opts`, so there
are two of them
- Fallback to using the same ones, but still keep them separate and
maintain them separately
- Main disadvantage of this approach is that the fallbacks for SSL and
IPv6 might need to happen twice if a separate connection string is not
being used - this can be improved by only populating the
`replication_opts`'s `connection_opts` right before starting the
replication client with the single `connection_opts` available if a
separate one was not provided.
- Make all tests (e.g. router tests) use a `pgBouncer` connection string
- Had to ensure connections are terminated explicitly at the end of the
test to avoid hanging connections from the pooler making the tests
really slow.

## TODO
- [x] Write integration test with `pgBouncer` as well that simply syncs
a shape with a pooled connection URL, although the unit tests already
cover that
- [ ] Avoid duplicate SSL and IPv6 fallbacks when separate query
connection string is not provided