chore: Improve transaction handling tests (#2442)
Closes https://github.com/electric-sql/electric/issues/2440

- Made the `ShapeLogCollector` test slightly more fuzzy, randomly tests
a few LSNs starting from 0 and generating successive values in the
64-bit range.
- I've made our tests and integration tests randomly reset the WAL to a
random point in the 64-bit range using `pg_resetwal` and a custom
`initdb` script - this doesn't apply to CI unit tests as it's a bit
tricky to add things to that service, but it should help with adding a
bit more fuzz to our elixir tests when ran locally.
- I've introduced a (slightly long) integration test that ensures that
replication works with 3 databases, starting at WAL `00000000/00000000`,
one at `7FFFFFFF/whatever`, and one at `FFFFFFFF/whatever`, so that it
spans the replication across the WAL range. It also has the benefit of
just testing the basic replication, since we didn't really have such a
straightforward integration test.
- Added test case that compares range of LSN values from start to the
64-bit end - bit of an overkill but just following the spirit of this
task

Fun fact this is what Postgres looks like when it reaches the end of the
available LSN range:

![image](https://github.com/user-attachments/assets/64f3eb6a-c2a8-41fa-ac7b-af3a98dd528d)
It definitely wasn't pretty - it just kept panicking and restarting.

I attempted to convert the `Lsn` struct to a struct whose structural
comparison still makes sense, and interestingly enough I got this
warning:

```bash
warning: comparison with structs found:
    left > right
given types:
    %Electric.Postgres.Lsn{offset: integer(), segment: integer()} > %Electric.Postgres.Lsn{
      offset: integer(),
      segment: integer()
    }
where "left" (context ExUnit.Assertions) was given the type:
    # type: %Electric.Postgres.Lsn{offset: integer(), segment: integer()}
    # from: test/electric/postgres/lsn_test.exs:18
    left = %Electric.Postgres.Lsn{offset: 0, segment: 5}
where "right" (context ExUnit.Assertions) was given the type:
    # type: %Electric.Postgres.Lsn{offset: integer(), segment: integer()}
    # from: test/electric/postgres/lsn_test.exs:18
    right = %Electric.Postgres.Lsn{offset: 0, segment: 6}
Comparison operators (>, <, >=, <=, min, and max) perform structural and not semantic comparison. Comparing with a struct won't give meaningful results. Structs that can be compared typically define a compare/2 function within their modules that can be used for semantic comparison.

typing violation found at:
│
18 │     assert %Lsn{segment: 5} > %Lsn{segment: 6}
│                             ~
│
└─ test/electric/postgres/lsn_test.exs:18:29: Electric.Postgres.LsnTest."test LSN implement `List.Chars` protocol"/1
```

So it seems that Elixir does indeed warn when you do structural
comparisons, as they don't make sense @balegas @robacourt . I suspect
that we just have to wait for the type system to evolve a bit more, as I
tried adding Dialyzer types to see if I would get this warning on the
relevant code and it compiled just fine.

Alternatively I can refactor the Lsn to be a tuple, which has stable,
predictable comparisons, and we can access the `segment` and `offset`
parts with explicit accessors, and generate them with a `.new` method. I
think this makes sense but I'll defer to more experienced Elixir devs.